refactor: Move ConnThread class to dedicated file
